requirements.txt
# ============================================================================
# ProductDataChallenge - Python Dependencies
# ============================================================================
# Core
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
pyyaml>=6.0

# ML & Feature Engineering
scikit-learn>=1.3.0
lightgbm>=4.0.0
mlxtend>=0.23.0
scipy>=1.11.0

# Text Processing
regex>=2023.0

# Explainability
shap>=0.43.0

# API Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Experiment Tracking
mlflow>=2.10.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Serialization
joblib>=1.3.0

# Dev & Testing
pytest>=7.4.0
httpx>=0.25.0

# Containerization support
gunicorn>=21.2.0
